chore: move the test function of `ChatGptEvaluationServiceIntegrationTest` to `ChatGptEvaluationServiceTest`
