+++ src/test/kotlin/org/elaastic/questions/assignment/sequence/interaction/chatGptEvaluation/ChatGptEvaluationServiceTest.kt
@@ -37,303 +37,4 @@ class ChatGptEvaluationServiceTest(
         // Precondition
         assertThat(chatGptEvaluationRepository.findAll(), `is`(empty()))
     }
-
-    @Test
-    fun `canHideGrading should return true if the user is the teacher of the sequence`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-        val teacher: User = response.interaction.owner
-
-        tGiven("A chatGPT evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                status = ChatGptEvaluationStatus.DONE.name,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = false,
-                removedByTeacher = false,
-            )
-        }.tThen("canHideGrading should return true") {
-            assertTrue(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, teacher))
-            it
-        }.tThen("canHideGrading should return false") {
-            val anotherUser = integrationTestingService.getTestStudent()
-            assertNotEquals(teacher, anotherUser)
-            assertFalse(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, anotherUser))
-        }
-    }
-
-    @Test
-    fun `canHideGrading should return false when the evaluation isn't DONE`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-        val teacher: User = response.interaction.owner
-
-        tGiven("A chatGPT evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = false,
-                removedByTeacher = false,
-            )
-        }.tWhen("The status isn't DONE") {
-            it.status = ChatGptEvaluationStatus.PENDING.name
-            assertNotEquals(ChatGptEvaluationStatus.DONE.name, it.status)
-            it
-        }.tThen("canHideGrading should return false") {
-            assertFalse(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, teacher))
-            it
-        }.tWhen("The status is DONE") {
-            it.status = ChatGptEvaluationStatus.DONE.name
-            assertEquals(ChatGptEvaluationStatus.DONE.name, it.status)
-            it
-        }.tThen("canHideGrading should return true") {
-            assertTrue(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, teacher))
-        }
-    }
-
-    @Test
-    fun `canHideGrading should return false when the user isn't the teacher of the sequence`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-        val teacher: User = response.interaction.owner
-        response.learner = integrationTestingService.getTestStudent()
-        responseRepository.save(response)
-
-        tGiven("A chatGPT evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                status = ChatGptEvaluationStatus.DONE.name,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = false,
-                removedByTeacher = false,
-            )
-        }.tThen("The user isn't the teacher of the sequence") {
-            val anotherUser = integrationTestingService.getTestStudent()
-            assertNotEquals(teacher, anotherUser)
-            assertFalse(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, anotherUser))
-            it
-        }.tThen("the learner of the response isn't the teacher of the sequence") {
-            assertNotEquals(teacher, response.learner)
-            assertFalse(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, response.learner))
-            it
-        }.tThen("canHideGrading should return true with the teacher") {
-            assertTrue(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, teacher))
-        }
-    }
-
-    @Test
-    fun `a teacher can hide a chatGPT evaluation in a sequence he own`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-        val teacher: User = response.interaction.owner
-
-        tGiven("A chatGPT evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                status = ChatGptEvaluationStatus.DONE.name,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = false,
-                removedByTeacher = false,
-            ).tWhen {
-                chatGptEvaluationRepository.save(it)
-                it
-            }
-        }.tWhen("The teacher hide the evaluation") {
-            assertDoesNotThrow {
-                chatGptEvaluationService.markAsHidden(it, teacher)
-            }
-            it
-        }.tThen("The evaluation is hidden") {
-            assertTrue(it.hiddenByTeacher)
-            it
-        }
-    }
-
-    @Test
-    fun `a student despit owning the response can't hide a chatGPT evaluation`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-        response.learner = integrationTestingService.getTestStudent()
-        responseRepository.save(response)
-        val student: User = response.learner
-
-        tGiven("A chatGPT evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                status = ChatGptEvaluationStatus.DONE.name,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = false,
-                removedByTeacher = false,
-            ).tWhen {
-                chatGptEvaluationRepository.save(it)
-                it
-            }
-        }.tWhen("The student try to hide the evaluation") {
-            assertFalse(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, student))
-            assertThrows(IllegalAccessException::class.java) {
-                chatGptEvaluationService.markAsHidden(it, student)
-            }
-            it
-        }.tThen("The evaluation is not hidden") {
-            assertFalse(it.hiddenByTeacher)
-            it
-        }
-    }
-
-    @Test
-    fun `a teacher can unhide a chatGPT evaluation in a sequence he own`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-        val teacher: User = response.interaction.owner
-
-        tGiven("A chatGPT hidden evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                status = ChatGptEvaluationStatus.DONE.name,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = true,
-                removedByTeacher = false,
-            ).tWhen {
-                chatGptEvaluationRepository.save(it)
-                it
-            }
-        }.tWhen("The teacher unhide the evaluation") {
-            assertTrue(it.hiddenByTeacher)
-            assertDoesNotThrow {
-                chatGptEvaluationService.markAsShown(it, teacher)
-            }
-            it
-        }.tThen("The evaluation is unhidden") {
-            assertFalse(it.hiddenByTeacher)
-            it
-        }
-    }
-
-    @Test
-    fun `a student despit owning the response can't unhide a chatGPT evaluation`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-        response.learner = integrationTestingService.getTestStudent()
-        responseRepository.save(response)
-        val student: User = response.learner
-
-        tGiven("A chatGPT hidden evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                status = ChatGptEvaluationStatus.DONE.name,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = true,
-                removedByTeacher = false,
-            ).tWhen {
-                chatGptEvaluationRepository.save(it)
-                it
-            }
-        }.tWhen("The student try to unhide the evaluation") {
-            assertTrue(it.hiddenByTeacher)
-            assertFalse(chatGptEvaluationService.canUpdateVisibilityEvaluation(it, student))
-            assertThrows(IllegalAccessException::class.java) {
-                chatGptEvaluationService.markAsShown(it, student)
-            }
-            it
-        }.tThen("The evaluation is not unhidden") {
-            assertTrue(it.hiddenByTeacher)
-            it
-        }
-    }
-
-    @Test
-    fun `nothing happend when we try to unhide an evalution that is visible`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-        val teacher: User = response.interaction.owner
-
-        tGiven("A chatGPT evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                status = ChatGptEvaluationStatus.DONE.name,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = false,
-                removedByTeacher = false,
-            ).tWhen {
-                chatGptEvaluationRepository.save(it)
-                it
-            }
-        }.tWhen("The teacher try to unhide the evaluation") {
-            assertFalse(it.hiddenByTeacher)
-            assertDoesNotThrow {
-                chatGptEvaluationService.markAsShown(it, teacher)
-            }
-            it
-        }.tThen("The evaluation is not unhidden") {
-            assertFalse(it.hiddenByTeacher)
-            it
-        }
-    }
-
-    @Test
-    fun `test of associateResponseToChatGPTEvaluationExistence`() {
-        // Given
-        val response = integrationTestingService.getAnyResponse()
-
-        var expected: Map<Long, Boolean> = mapOf(response.id!! to false)
-        assertEquals(
-            expected,
-            chatGptEvaluationService.associateResponseToChatGPTEvaluationExistence(listOf(response.id))
-        )
-
-        tGiven("A chatGPT evaluation") {
-            ChatGptEvaluation(
-                response = response,
-                annotation = "annotation",
-                grade = null,
-                status = ChatGptEvaluationStatus.DONE.name,
-                reportReasons = null,
-                reportComment = null,
-                utilityGrade = null,
-                hiddenByTeacher = false,
-                removedByTeacher = false,
-            ).tWhen {
-                chatGptEvaluationRepository.save(it)
-                it
-            }
-        }.tWhen("The teacher try to unhide the evaluation") {
-            expected = mapOf(response.id!! to true)
-            assertEquals(
-                expected,
-                chatGptEvaluationService.associateResponseToChatGPTEvaluationExistence(listOf(response.id))
-            )
-        }
-    }
 }
